tools/geo_space/requirements.txt
python-docx>=1.0
streamlit>=1.30
pandas>=2.0
pillow>=10.0
matplotlib>=3.8
